Bug 1298756 - Use uint8_t/uint16_t as underlying type for enums from webidl. r=bz
MozReview-Commit-ID: 3uS9s5ZrPFd

+++ dom/base/nsGlobalWindow.h
@@ -114,7 +114,7 @@ class CustomElementRegistry;
 class DocGroup;
 class External;
 class Function;
-enum class ImageBitmapFormat : uint32_t;
+enum class ImageBitmapFormat : uint8_t;
 class IdleRequest;
 class IdleRequestCallback;
 class Location;
